# marginal improvement: atomic byte-counts